Update Color.cs (#3924)
* Update Color.cs

Added typecast in "public static Color FromARGB(int argb)" to prevent an infinite loop.

* Update CHANGELOG.md

---------

+++ src/ScottPlot5/ScottPlot5/Primitives/Color.cs
@@ -111,7 +111,7 @@ public readonly struct Color
 
     public static Color FromARGB(int argb)
     {
-        return FromARGB(argb);
+        return FromARGB((uint)argb);
     }
 
     public static Color FromARGB(uint argb)
